Handle Windows and remote buffers using default formatters
Various fixes to ensure that the default formatters work over Tramp
and on Windows:

1. Replace the apheleia-npx command with npx when the buffer is remote
or running on Windows. This uses the built-in npx handling to lookup
the binary, so npx never gets called. Since the apheleia-npx script
isn't available on remote hosts, it makes sense to remove this. The
same is done for Windows to capture the case of using WSL, in which
case calling bash will enter WSL and so the script cannot be accessed
using the Windows path.

2. Use local file name when returning binary path.

3. Don't use executable-find if the binary is already an absolute path.

4. Pass the -l option to sh to ensure that it loads the user's profile.